+++ NEWS.rst
@@ -12,6 +12,7 @@ Version History
 * Vendored version of zstd updated to 1.1.2.
 * Documentation improvements.
 * Introduce a bench.py script for performing (crude) benchmarks.
+* ZSTD_CCtx instances are now reused across multiple compress() operations.
 
 0.5.2 (released 2016-11-12)
 ---------------------------